fix:moved  function to controllers
